added product catalog implementation and resolved issues with hibernate mapping.

+++ eStoreRest/src/main/java/com/ss/estore/dao/StoreDAO.java
@@ -16,5 +16,5 @@ public interface StoreDAO {
 	public void delete(long storeId);
 	public List<Store> list();
 
-	public Store fetch(int storeId);
+	public Store fetch(long storeId);
 }
